requirements.txt
Flask
flasgger
azure-storage-blob
azure-ai-documentintelligence
werkzeug
